200212_kijoon_결재 수정, 호텔정보 수정, 수리내역 추가

See merge request ghks8198/hotelscompile!126

+++ src/main/java/com/kh/hotels/mngApproval/model/vo/OrderRequest.java
@@ -35,13 +35,13 @@ public class OrderRequest {
 	private String mfg;
 	private int vosprice;
 	private int price;
-	private int docno;
+	private Long docno;
 	private String deptname;
 	private String sinfo;
 	private int smno;
 	private int mno;
 	private String stfId;
-	private int purDocuno;
+	private Long purDocuno;
 	private String rsvReqDate;
 	
 }
